flows: move flow diagram logic to separate file
